* Added CommandCollection interface to facilitate adding commands from class files (outside of CommandProcessor) that contain multiple command definitions. CommandProcessor does not have to be aware of what specific commands are contained in the individual CommandCollection implementations, so CommandCollection implementations may now be extended without modifying CommandProcessor.  * Fixed overflow in InventoryTools.getAmountOfItem. Now returns an int.  * Added !take command for taking an item from a player.  * Fixed bugs related to spawn points when exiting through the Free Market portal to the default map, Perion, when the saved location data is missing.  * Allow warp to a map if the WZs for a portal have an invalid destination.  * Less sensitive fast HP/MP regeneration detection.  * Removed infraction triggers in AP and SP allocation because latency or tubi enabled clients may fire them.  * Identified cause of Energy Charge effect showing in place of Dash.

+++ src/argonms/game/field/entity/Mob.java
@@ -210,7 +210,7 @@ public class Mob extends AbstractEntity {
 		}
 		for (Integer itemId : stats.getItemsToTake()) {
 			Inventory.InventoryType type = InventoryTools.getCategory(itemId);
-			short quantity = InventoryTools.getAmountOfItem(killer.getInventory(type), itemId);
+			int quantity = InventoryTools.getAmountOfItem(killer.getInventory(type), itemId);
 			if (type == Inventory.InventoryType.EQUIP)
 				quantity += InventoryTools.getAmountOfItem(killer.getInventory(Inventory.InventoryType.EQUIPPED), itemId);
 			if (quantity > 0) {
